+++ chapter09/src/test/java/com/seaofnodes/simple/Chapter08Test.java
@@ -4,7 +4,6 @@ import com.seaofnodes.simple.node.*;
 import com.seaofnodes.simple.type.TypeInteger;
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.Ignore;
 
 import static org.junit.Assert.*;
