+++ chapter10/src/test/java/com/seaofnodes/simple/Chapter08Test.java
@@ -5,7 +5,6 @@ import com.seaofnodes.simple.node.*;
 import com.seaofnodes.simple.type.TypeInteger;
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.Ignore;
 
 import static org.junit.Assert.*;
