+++ chapter11/src/test/java/com/seaofnodes/simple/Chapter03Test.java
@@ -2,9 +2,7 @@ package com.seaofnodes.simple;
 
 import com.seaofnodes.simple.node.*;
 import com.seaofnodes.simple.type.TypeInteger;
-import org.junit.Assert;
 import org.junit.Test;
-import org.junit.Ignore;
 
 import static org.junit.Assert.*;
 
@@ -58,7 +56,7 @@ public class Chapter03Test {
     }
 
     @Test
-    public void testBad6() {
+    public void testBad1() {
         try {
             new Parser("int a=1; int b=2; int c=0; { int b=3; c=a+b;").parse();
             fail();
