+++ chapter11/src/test/java/com/seaofnodes/simple/Chapter05Test.java
@@ -2,9 +2,7 @@ package com.seaofnodes.simple;
 
 import com.seaofnodes.simple.node.*;
 import com.seaofnodes.simple.type.TypeInteger;
-import org.junit.Assert;
 import org.junit.Test;
-import org.junit.Ignore;
 
 import static org.junit.Assert.*;
 
@@ -203,22 +201,12 @@ return a;""");
     }
 
     @Test
-        public void testKeyword3() {
+    public void testKeyword3() {
         try {
             new Parser("int a=1; ififif(arg)inta=2;return a;").parse();
             fail();
         } catch( RuntimeException e ) {
             assertEquals("Undefined name 'ififif'",e.getMessage());
-        }
-    }
-
-    @Test
-    public void testBad7() {
-        try {
-            new Parser("return 1;}").parse();
-            fail();
-        } catch( RuntimeException e ) {
-            assertEquals("Syntax error, unexpected }",e.getMessage());
         }
     }
